Move DataTypeConstants out of store file and into constants
