+++ src/pages/resultsView/plots/PlotsTab.tsx
@@ -9,11 +9,8 @@ import {
 } from 'mobx';
 import { Observer, observer } from 'mobx-react';
 import './styles.scss';
-import {
-    DataTypeConstants,
-    ResultsViewPageStore,
-} from '../ResultsViewPageStore';
-import { AlterationTypeConstants } from 'shared/constants';
+import { ResultsViewPageStore } from '../ResultsViewPageStore';
+import { AlterationTypeConstants, DataTypeConstants } from 'shared/constants';
 import { Button, FormControl } from 'react-bootstrap';
 import ReactSelect from 'react-select1';
 import Select from 'react-select';
